Float replaced with double data type

+++ apps/java/libraries/case-engine/src/main/java/com/wks/caseengine/dto/AOPReportDTO.java
@@ -17,18 +17,18 @@ public class AOPReportDTO {
     private String norm;
     private String particulars;
 
-    private Float april;
-    private Float may;
-    private Float june;
-    private Float july;
-    private Float august;
-    private Float september;
-    private Float october;
-    private Float november;
-    private Float december;
-    private Float january;
-    private Float february;
-    private Float march;
-    private Float total;
+    private Double april;
+    private Double may;
+    private Double june;
+    private Double july;
+    private Double august;
+    private Double september;
+    private Double october;
+    private Double november;
+    private Double december;
+    private Double january;
+    private Double february;
+    private Double march;
+    private Double total;
 
 }
