+++ apps/java/libraries/case-engine/src/main/java/com/wks/caseengine/dto/BusinessDemandDTO.java
@@ -22,26 +22,26 @@ public class BusinessDemandDTO {
     private String normParameterId;
     private String catalystId;
     private String type;
-    private Float jan;
-    private Float feb;
-    private Float march;
-    private Float april;
-    private Float may;
-    private Float june;
-    private Float july;
-    private Float aug;
-    private Float sep;
-    private Float oct;
-    private Float nov;
-    private Float dec;
+    private Double jan;
+    private Double feb;
+    private Double march;
+    private Double april;
+    private Double may;
+    private Double june;
+    private Double july;
+    private Double aug;
+    private Double sep;
+    private Double oct;
+    private Double nov;
+    private Double dec;
     private String year;
     private String plantFkId;
-    private Float TPH;
-    private Float avgTPH;
+    private Double TPH;
+    private Double avgTPH;
     private String UOM;
 
-    public Float getMonthValue(Integer month) {
-        Float value = 0.0F;
+    public Double getMonthValue(Integer month) {
+        Double value = 0.0;
         switch (month) {
             case 1:
                 value = getJan();
@@ -81,7 +81,7 @@ public class BusinessDemandDTO {
                 break;
 
             default:
-                value = 0.0F; // Default case if no match is found
+                value = 0.0; // Default case if no match is found
         }
         return value;
     }
