+++ apps/java/libraries/case-engine/src/main/java/com/wks/caseengine/service/MaintenanceCalculatedDataServiceImpl.java
@@ -48,18 +48,18 @@ public class MaintenanceCalculatedDataServiceImpl implements MaintenanceCalculat
 			for (Object[] row : list) {
 				MaintenanceDetailsDTO dto = new MaintenanceDetailsDTO();
 				dto.setName(row[2] != null ? row[2].toString() : null);
-				dto.setJan(row[3] != null ? Float.valueOf(row[3].toString()) : null);
-				dto.setFeb(row[4] != null ? Float.valueOf(row[4].toString()) : null);
-				dto.setMar(row[5] != null ? Float.valueOf(row[5].toString()) : null);
-				dto.setApril(row[6] != null ? Float.valueOf(row[6].toString()) : null);
-				dto.setMay(row[7] != null ? Float.valueOf(row[7].toString()) : null);
-				dto.setJune(row[8] != null ? Float.valueOf(row[8].toString()) : null);
-				dto.setJuly(row[9] != null ? Float.valueOf(row[9].toString()) : null);
-				dto.setAug(row[10] != null ? Float.valueOf(row[10].toString()) : null);
-				dto.setSep(row[11] != null ? Float.valueOf(row[11].toString()) : null);
-				dto.setOct(row[12] != null ? Float.valueOf(row[12].toString()) : null);
-				dto.setNov(row[13] != null ? Float.valueOf(row[13].toString()) : null);
-				dto.setDec(row[14] != null ? Float.valueOf(row[14].toString()) : null);
+				dto.setJan(row[3] != null ? Double.valueOf(row[3].toString()) : null);
+				dto.setFeb(row[4] != null ? Double.valueOf(row[4].toString()) : null);
+				dto.setMar(row[5] != null ? Double.valueOf(row[5].toString()) : null);
+				dto.setApril(row[6] != null ? Double.valueOf(row[6].toString()) : null);
+				dto.setMay(row[7] != null ? Double.valueOf(row[7].toString()) : null);
+				dto.setJune(row[8] != null ? Double.valueOf(row[8].toString()) : null);
+				dto.setJuly(row[9] != null ? Double.valueOf(row[9].toString()) : null);
+				dto.setAug(row[10] != null ? Double.valueOf(row[10].toString()) : null);
+				dto.setSep(row[11] != null ? Double.valueOf(row[11].toString()) : null);
+				dto.setOct(row[12] != null ? Double.valueOf(row[12].toString()) : null);
+				dto.setNov(row[13] != null ? Double.valueOf(row[13].toString()) : null);
+				dto.setDec(row[14] != null ? Double.valueOf(row[14].toString()) : null);
 				maintenanceDetailsDTOList.add(dto);
 			}
